flask-server/requirements.txt
flask==1.1.2 
flask_json==0.3.4
pwjwt==2.0.1
flask-sqlalchemy==2.4.4
